perf: isolate AI vendor model types
